Script to auto deploy in the ClusterPi

+++ deploy.sh
@@ -3,7 +3,8 @@
 # GitHub repository URL
 REPO_URL=https://github.com/cycl0ps22/ClusterPi.git
 # Directory to clone the repository
-CLONE_DIR=/tmp/ClusterPi
+CLONE_DIR_BASE=/tmp
+CLONE_DIR=$CLONE_DIR_BASE/ClusterPi
 
 # Function to deploy changed YAML files
 deploy_changes() {
@@ -11,6 +12,7 @@ deploy_changes() {
     if [ ${#changed_files[@]} -eq 0 ]; then
         echo "No changes detected in YAML files."
     else
+        echo "Changes detected !!!!!!!!!!"
         echo "Deploying changed YAML files:"
         for file in "${changed_files[@]}"; do
             echo "- $file"
@@ -29,11 +31,18 @@ detect_changes() {
 }
 
 # Main script execution starts here
-echo "Cloning the GitHub repository..."
-git clone "$REPO_URL" "$CLONE_DIR" || { echo "Error: Unable to clone the repository."; exit 1; }
-
-echo "Checking for changes in YAML files..."
-detect_changes || echo "No changes detected in YAML files."
-
-echo "Removing the cloned repository directory..."
-rm -rf "$CLONE_DIR"
+if [ ! -d "$CLONE_DIR" ]; then
+    echo "First run: Cloning the GitHub repository..."
+    git clone "$REPO_URL" "$CLONE_DIR" || { echo "Error: Unable to clone the repository."; exit 1; }
+    echo "Checking for changes in YAML files..."
+    detect_changes || echo "No changes detected in YAML files."
+else
+    echo "Second run: Cloning the GitHub repository for comparison..."
+    CLONE_DIR_NEW="$CLONE_DIR_BASE/ClusterPi_new"
+    git clone "$REPO_URL" "$CLONE_DIR_NEW" || { echo "Error: Unable to clone the repository."; exit 1; }
+    echo "Checking for changes in YAML files..."
+    detect_changes || echo "No changes detected in YAML files."
+    echo "Removing the old cloned repository directory..."
+    rm -rf "$CLONE_DIR"
+    mv "$CLONE_DIR_NEW" "$CLONE_DIR"
+fi
